Optimize updates of status bar during search.
Optimization is based on examining file size and time of the last status
update. This is synchronous and is based on assumption that relatively
small files are quickly processed even on slow hardware so if refresh
timeout expires during their processing, the delay couldn't be noticed
by human. Therefore only rather big files have to be reported
unconditionally, and others - only if there is a match in this file or
if timeout expires.

Signed-off-by: Andrew Borodin <aborodin@vmail.ru>